Suppress warnings for suspicious use of "enum" constants.
In each case, the code is using values of enums in valid ways that
GCC 4.7 does not expect, so use a cast to suppress the warnings.

+++ src/language/data-io/placement-parser.c
@@ -270,7 +270,7 @@ bool
 execute_placement_format (const struct fmt_spec *format,
                           int *record, int *column)
 {
-  switch (format->type)
+  switch ((int) format->type)
     {
     case PRS_TYPE_X:
       *column += format->w;
